requirements.txt
absl-py==2.0.0
accelerate==0.18.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
anthropic==0.25.6
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bertviz==1.4.0
bitsandbytes==0.38.1
bleach==6.1.0
boto3==1.28.70
botocore==1.31.70
braceexpand==0.1.7
Brotli==1.0.9
build==1.2.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
chatarena==0.1.18
chroma-hnswlib==0.7.3
chromadb==0.4.24
cityscapesScripts==2.2.1
click==8.1.7
clip==1.0
cloudpickle==3.0.0
cmake==3.27.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.6.4
datasets==2.14.6
debugpy==1.8.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.9.5
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distro==1.8.0
diversipy==0.8
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.6.0
einops-exts==0.0.4
evalplus==0.2.0
exceptiongroup==1.1.3
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.103.0
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.12.3
flash-attn==2.1.1
flatbuffers==24.3.25
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-generativeai==0.5.2
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gradio==3.35.2
gradio_client==0.2.9
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
gymnasium==0.29.1
h11==0.14.0
h5py==3.8.0
highway-env==1.8.2
hjson==3.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.0
importlib_resources==6.4.0
inflect==6.0.2
iopath==0.1.10
ipykernel==6.22.0
ipython==8.11.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.1
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.10
keras==2.11.0
kiwisolver==1.4.5
kornia==0.6.9
kubernetes==29.0.0
langchain==0.1.16
langchain-anthropic==0.1.11
langchain-community==0.0.34
langchain-core==0.1.45
langchain-experimental==0.0.57
langchain-google-genai==1.0.3
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.47
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.10.1
linkify-it-py==2.0.2
lit==16.0.6
llava==1.1.3
maplm==1.0.0
Markdown==3.5
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.6.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
num2words==0.5.12
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.3
open-clip-torch==2.22.0
open-flamingo==2.0.1
openai==1.30.1
opencv-python==4.8.0.74
opencv-python-headless==4.5.5.64
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
peft==0.4.0
pettingzoo==1.24.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3
platformdirs==4.2.2
portalocker==2.8.2
posthog==3.5.0
prettytable==3.6.0
proglog==0.1.10
progressbar2==4.4.2
prometheus_client==0.20.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.1.1
pycocotools==2.0.6
pycparser==2.22
pydantic==1.10.12
pydub==0.25.1
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.0.0
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
python-utils==3.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==26.0.3
qd==0.8.9
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.30.2
regex==2022.10.31
replicate==0.25.2
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rpds-py==0.10.0
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
setuptools==65.6.3
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tempdir==0.7.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.2.0
tifffile==2024.5.3
tiktoken==0.6.0
timm==0.6.12
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.3.1
torchvision==0.14.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.3
transformers==4.41.0.dev0
triton==2.0.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing==3.7.4.3
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
visual-genome==1.1.1
wandb==0.13.9
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.8
webcolors==1.13
webdataset==0.2.62
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.0
wget==3.2
wheel==0.38.4
widgetsnbextension==4.0.10
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.18.1
